ML Fine-tuning

+++ processing/audio.py
@@ -65,8 +65,8 @@ class Audio:
         return NotImplemented
 
     def save(self, path):
-        data = np.array(self.time_series, dtype=np.float32)
-        scipy.io.wavfile.write(path, self.__sampling_rate, data)
+        data = np.array(self.time_series)
+        scipy.io.wavfile.write(path, self.__sampling_rate, data.astype(np.int16))
 
     def mel_spectrogram(self):
         data = np.array(self.time_series, dtype=np.float32)
